# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestExperience =
  (typeof RegisterRequestExperience)[keyof typeof RegisterRequestExperience];

export const RegisterRequestExperience = {
  beginner: "beginner",
  experienced: "experienced",
} as const;

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  experience: RegisterRequestExperience;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserKycStatus = (typeof UserKycStatus)[keyof typeof UserKycStatus];

export const UserKycStatus = {
  unverified: "unverified",
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  suspended: "suspended",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  experience: string;
  usdBalance: number;
  kycStatus: UserKycStatus;
  role: UserRole;
  status: UserStatus;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface AdminStats {
  totalUsers: number;
  totalAdmins: number;
  verifiedUsers: number;
  suspendedUsers: number;
  totalUsdBalance: number;
  totalCryptoValue: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
  completedTransactions: number;
  totalVolumeUsd: number;
}

export interface AdminTransaction {
  id: number;
  userId: number;
  userEmail: string;
  userName: string;
  type: string;
  coin?: string | null;
  symbol?: string | null;
  amount?: number | null;
  usdAmount: number;
  price?: number | null;
  status: string;
  createdAt: string;
}

export type AdminUserUpdateKycStatus =
  (typeof AdminUserUpdateKycStatus)[keyof typeof AdminUserUpdateKycStatus];

export const AdminUserUpdateKycStatus = {
  unverified: "unverified",
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export type AdminUserUpdateRole =
  (typeof AdminUserUpdateRole)[keyof typeof AdminUserUpdateRole];

export const AdminUserUpdateRole = {
  user: "user",
  admin: "admin",
} as const;

export type AdminUserUpdateStatus =
  (typeof AdminUserUpdateStatus)[keyof typeof AdminUserUpdateStatus];

export const AdminUserUpdateStatus = {
  active: "active",
  suspended: "suspended",
} as const;

export interface AdminUserUpdate {
  usdBalance?: number;
  kycStatus?: AdminUserUpdateKycStatus;
  role?: AdminUserUpdateRole;
  status?: AdminUserUpdateStatus;
  adjustBalance?: number;
  adjustReason?: string;
}

export interface KycSubmitRequest {
  fullName: string;
  dateOfBirth: string;
  country: string;
  idNumber: string;
}

export interface Holding {
  coin: string;
  symbol: string;
  amount: number;
  avgBuyPrice: number;
  currentPrice: number;
  currentValue: number;
  pnl: number;
  pnlPercent: number;
}

export interface Portfolio {
  usdBalance: number;
  totalValue: number;
  holdings: Holding[];
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  buy: "buy",
  sell: "sell",
  deposit: "deposit",
  withdraw: "withdraw",
  convert: "convert",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  completed: "completed",
  pending: "pending",
  failed: "failed",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  coin?: string | null;
  symbol?: string | null;
  amount?: number | null;
  usdAmount: number;
  price?: number | null;
  status: TransactionStatus;
  createdAt: string;
}

export interface TradeRequest {
  symbol: string;
  usdAmount: number;
}

export interface DepositRequest {
  /** When symbol is provided, this is the coin amount (e.g. 0.5 BTC). Otherwise it is the USD amount credited to fiat balance. */
  amount: number;
  method: string;
  /** Coin symbol to credit (e.g. BTC, USDT). If omitted, deposit credits the USD fiat balance. */
  symbol?: string;
  address?: string;
}

export interface ConvertRequest {
  fromSymbol: string;
  toSymbol: string;
  /** Amount of fromSymbol coin to convert. */
  fromAmount: number;
}

export interface ConvertResponse {
  fromSymbol: string;
  toSymbol: string;
  fromAmount: number;
  toAmount: number;
  rate: number;
  usdValue: number;
}

export interface WithdrawRequest {
  amount: number;
  method: string;
  address?: string;
}

export interface CoinPrice {
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  changePercent24h: number;
  volume24h: number;
  marketCap: number;
  icon: string;
}

export type GetAdminUsersParams = {
  search?: string;
};

export type GetAdminTransactionsParams = {
  status?: string;
  type?: string;
  userId?: number;
};