# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoX Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/kyc/verify:
    post:
      operationId: submitKyc
      tags: [auth]
      summary: Submit KYC verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitRequest"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /transactions/deposit:
    post:
      operationId: deposit
      tags: [transactions]
      summary: Deposit funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats (admin only)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: List all transactions (admin only)
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTransaction"
  /admin/transactions/{id}/approve:
    post:
      operationId: approveAdminTransaction
      tags: [admin]
      summary: Approve a pending transaction (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransaction"
  /admin/transactions/{id}/reject:
    post:
      operationId: rejectAdminTransaction
      tags: [admin]
      summary: Reject a pending transaction (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransaction"
...
  /transactions/withdraw:
    post:
      operationId: withdraw
      tags: [transactions]
      summary: Withdraw funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        experience:
          type: string
          enum: [beginner, experienced]
      required:
        - email
        - password
        - name
        - experience
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        experience:
          type: string
        usdBalance:
          type: number
        kycStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - experience
        - usdBalance
        - kycStatus
        - role
        - status
        - createdAt
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalAdmins:
          type: integer
        verifiedUsers:
          type: integer
        suspendedUsers:
          type: integer
        totalUsdBalance:
          type: number
        totalCryptoValue:
          type: number
        pendingDeposits:
          type: integer
        pendingWithdrawals:
          type: integer
        completedTransactions:
          type: integer
        totalVolumeUsd:
          type: number
      required:
        - totalUsers
        - totalAdmins
        - verifiedUsers
        - suspendedUsers
        - totalUsdBalance
        - totalCryptoValue
        - pendingDeposits
        - pendingWithdrawals
        - completedTransactions
        - totalVolumeUsd
    AdminTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userName:
          type: string
        type:
          type: string
        coin:
          type: string
          nullable: true
        symbol:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        usdAmount:
          type: number
        price:
          type: number
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - userEmail
        - userName
        - type
        - usdAmount
        - status
        - createdAt
    AdminUserUpdate:
      type: object
      properties:
        usdBalance:
          type: number
        kycStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended]
        adjustBalance:
          type: number
        adjustReason:
          type: string
    KycSubmitRequest:
      type: object
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        country:
          type: string
        idNumber:
          type: string
      required:
        - fullName
        - dateOfBirth
        - country
        - idNumber
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [buy, sell, deposit, withdraw, convert]
        coin:
          type: string
          nullable: true
        symbol:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        usdAmount:
          type: number
        price:
          type: number
          nullable: true
        status:
          type: string
          enum: [completed, pending, failed]
        createdAt:
          type: string
      required:
        - id
        - type
        - usdAmount
        - status
        - createdAt
...
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
          description: When symbol is provided, this is the coin amount (e.g. 0.5 BTC). Otherwise it is the USD amount credited to fiat balance.
        method:
          type: string
        symbol:
          type: string
          description: Coin symbol to credit (e.g. BTC, USDT). If omitted, deposit credits the USD fiat balance.
        address:
          type: string
      required:
        - amount
        - method
...
    WithdrawRequest:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
        address:
          type: string
      required:
        - amount
        - method
...