# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalAdmins: number;
  verifiedUsers: number;
  suspendedUsers: number;
  totalUsdBalance: number;
  totalCryptoValue: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
  completedTransactions: number;
  totalVolumeUsd: number;
}