# lib/api-zod/src/generated/types/adminTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminTransaction {
  id: number;
  userId: number;
  userEmail: string;
  userName: string;
  type: string;
  coin?: string | null;
  symbol?: string | null;
  amount?: number | null;
  usdAmount: number;
  price?: number | null;
  status: string;
  createdAt: string;
}