# lib/api-zod/src/generated/types/adminUserUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserUpdateKycStatus } from "./adminUserUpdateKycStatus";
import type { AdminUserUpdateRole } from "./adminUserUpdateRole";
import type { AdminUserUpdateStatus } from "./adminUserUpdateStatus";

export interface AdminUserUpdate {
  usdBalance?: number;
  kycStatus?: AdminUserUpdateKycStatus;
  role?: AdminUserUpdateRole;
  status?: AdminUserUpdateStatus;
  adjustBalance?: number;
  adjustReason?: string;
}